--- parser/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .model import Model
-
-__all__ = ['Model']
