--- parser/cmds/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .evaluate import Evaluate
-from .predict import Predict
-from .train import Train
-
-__all__ = ['Evaluate', 'Predict', 'Train']
