--- requirements.txt
@@ -1,1 +1,0 @@
-transformers == 2.1.1
