+++ supar/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from .biaffine_parser import BiaffineParser
+
+__all__ = ['BiaffineParser']
