+++ supar/modules/bilstm.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
 
-from parser.modules.dropout import SharedDropout
-
 import torch
 import torch.nn as nn
+from supar.modules.dropout import SharedDropout
 from torch.nn.modules.rnn import apply_permutation
 from torch.nn.utils.rnn import PackedSequence
